+++ projects/myrmidon/cadmus-profile-editor/src/public-api.ts
@@ -2,6 +2,7 @@
  * Public API Surface of cadmus-profile-editor
  */
 
+export * from './lib/services/facet-definition-validator.service';
 export * from './lib/components/facet-definition-editor/facet-definition-editor.component';
 export * from './lib/components/facet-definition-list-editor/facet-definition-list-editor.component';
 export * from './lib/components/part-definition-editor/part-definition-editor.component';